change VelocyPack comparison functions
This PR renames the existing method velocypack::Slice::equals() to
velocypack::Slice::binaryEquals, in order to make clear that the
comparison is done on the binary level. The binary comparison function
may consider two Slices unequal if they have a different binary
representation but logically contain the same data, e.g. when using
an Array with an index table vs. using an Array without an index table.

The behavior of equals (now: binaryEquals) is not changed by this PR,
it is just that the naming is more intuitive.
In order to avoid unintentional binary comparisons of Slices, the
operators `==` and `!=` of velocypack::Slice have also been removed.

To compare Slices, it is still possible to call the new/renamed
binaryEquals method. Removing `operator==` of velocypack::Slice also
leads to Slices not being usable as keys in hashed containers (e.g.
unordered_map, unordered_set) without additional effort.

Helper structs for binary comparisons are added by this PR though:

* arangodb::velocypack::BinaryCompare::Hash for hashing
* arangodb::velocypack::BinaryCompare::Equal for equality testing

Instances of these objects can be used together with unordered
containers as follows:

    std::unordered_set<
      arangodb::velocypack::Slice,
      arangodb::velocypack::BinaryCompare::Hash,
      arangodb::velocypack::BinaryCompare::Equal
    >