remove default parameter from ObjectIterator ctor
the second argument of the `ObjectIterator` ctor determines whether the
attributes in the object should be iterated in sorted order (sorted by
attribute names) or be iterated in storage order.
that parameter previously defaulted to `false`, meaning that sorted
iteration was used by default. sorted iteration may come with a
performance overhead, so it is better to make the choice explicit.
this change removes the default value for the parameter, so the user now
must explicitly set the iteration order to either
- `false`: iterate in sorted order (same as default value before)
- `true`: iterate in storage order